update new modification.

+++ src/components/application_manager/include/application_manager/policies/policy_handler.h
@@ -314,10 +314,10 @@ class PolicyHandler :
   /**
    * Returns heart beat timeout
    * @param app_id application id
-   * @return if timeout was set then value in seconds greater zero
+   * @return if timeout was set then value in milliseconds greater zero
    * otherwise heart beat for specific application isn't set
    */
-  uint16_t HeartBeatTimeout(const std::string& app_id) const;
+  uint32_t HeartBeatTimeout(const std::string& app_id) const;
 
   /**
    * @brief Returns URL for querying list of remote apps
